tests/wpt/tests/html/canvas/offscreen/text/2d.text.draw.align.start.rtl.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.draw.align.start.rtl</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.draw.align.start.rtl</h1>
<p class="desc">textAlign start with rtl is the right edge</p>


<script>
var t = async_test("textAlign start with rtl is the right edge");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var f = new FontFace("CanvasTest", "url('/fonts/CanvasTest.ttf')");
  let fonts = (self.fonts ? self.fonts : document.fonts);
  f.load();
  fonts.add(f);
  fonts.ready.then(function() {
      ctx.font = '50px CanvasTest';
      ctx.direction = 'rtl';
      ctx.fillStyle = '#f00';
      ctx.fillRect(0, 0, 100, 50);
      ctx.fillStyle = '#0f0';
      ctx.textAlign = 'start';
      ctx.fillText('DD', 100, 37.5);
      _assertPixelApprox(canvas, 5,5, 0,255,0,255, 2);
      _assertPixelApprox(canvas, 95,5, 0,255,0,255, 2);
      _assertPixelApprox(canvas, 25,25, 0,255,0,255, 2);
      _assertPixelApprox(canvas, 75,25, 0,255,0,255, 2);
      _assertPixelApprox(canvas, 5,45, 0,255,0,255, 2);
      _assertPixelApprox(canvas, 95,45, 0,255,0,255, 2);
    }).then(t_pass, t_fail);

});
</script>
